@@ -1,64 +1,73 @@
 {
-	"name": "taro-msparis",
-	"version": "1.0.0",
-	"private": true,
-	"description": "A taro project",
-	"main": "index.js",
-	"scripts": {
-		"dev:weapp": "npm run build:weapp -- --watch",
-		"dev:alipay": "npm run build:alipay -- --watch",
-		"dev:swan": "npm run build:swan -- --watch",
-		"dev:tt": "npm run build:tt -- --watch",
-		"dev:h5": "npm run build:h5 -- --watch",
-		"dev:rn": "npm run build:rn -- --watch",
-		"build:weapp": "taro build --type weapp",
-		"build:alipay": "taro build --type alipay",
-		"build:swan": "taro build --type swan",
-		"build:tt": "taro build --type tt",
-		"build:h5": "taro build --type h5",
-		"build:rn": "taro build --type rn",
-		"tep": "node template"
-	},
-	"author": "cai.bowen <easytuan@foxmail.com>",
-	"license": "MIT",
-	"dependencies": {
-		"@tarojs/async-await": "1.2.2",
-		"@tarojs/components": "1.2.2",
-		"@tarojs/redux": "1.2.2",
-		"@tarojs/redux-h5": "1.2.2",
-		"@tarojs/rn-runner": "1.2.2",
-		"@tarojs/router": "1.2.2",
-		"@tarojs/taro": "1.2.2",
-		"@tarojs/taro-alipay": "1.2.2",
-		"@tarojs/taro-h5": "1.2.2",
-		"@tarojs/taro-swan": "1.2.2",
-		"@tarojs/taro-tt": "^1.2.2",
-		"@tarojs/taro-weapp": "1.2.2",
-		"dva-core": "^1.4.0",
-		"dva-loading": "^2.0.5",
-		"nervjs": "^1.3.0",
-		"redux": "^4.0.0",
-		"redux-logger": "^3.0.6",
-		"redux-thunk": "^2.3.0"
-	},
-	"devDependencies": {
-		"@tarojs/plugin-babel": "1.2.2",
-		"@tarojs/plugin-csso": "1.2.2",
-		"@tarojs/plugin-sass": "1.2.2",
-		"@tarojs/plugin-uglifyjs": "1.2.2",
-		"@tarojs/webpack-runner": "1.2.2",
-		"@types/react": "^16.4.8",
-		"@types/webpack-env": "^1.13.6",
-		"babel-eslint": "^8.2.3",
-		"babel-plugin-transform-class-properties": "^6.24.1",
-		"babel-plugin-transform-decorators-legacy": "^1.3.4",
-		"babel-plugin-transform-jsx-stylesheet": "^0.6.5",
-		"babel-plugin-transform-object-rest-spread": "^6.26.0",
-		"babel-preset-env": "^1.6.1",
-		"eslint": "^4.19.1",
-		"eslint-config-taro": "1.2.2",
-		"eslint-plugin-import": "^2.12.0",
-		"eslint-plugin-react": "^7.8.2",
-		"eslint-plugin-taro": "1.2.2"
-	}
+  "name": "mymarket",
+  "version": "1.0.0",
+  "private": true,
+  "description": "多端适用的电子商城",
+  "templateInfo": {
+    "name": "redux",
+    "typescript": true,
+    "css": "sass"
+  },
+  "scripts": {
+    "build:weapp": "taro build --type weapp",
+    "build:swan": "taro build --type swan",
+    "build:alipay": "taro build --type alipay",
+    "build:tt": "taro build --type tt",
+    "build:h5": "taro build --type h5",
+    "build:rn": "taro build --type rn",
+    "build:qq": "taro build --type qq",
+    "build:jd": "taro build --type jd",
+    "build:quickapp": "taro build --type quickapp",
+    "dev:weapp": "npm run build:weapp -- --watch",
+    "dev:swan": "npm run build:swan -- --watch",
+    "dev:alipay": "npm run build:alipay -- --watch",
+    "dev:tt": "npm run build:tt -- --watch",
+    "dev:h5": "npm run build:h5 -- --watch",
+    "dev:rn": "npm run build:rn -- --watch",
+    "dev:qq": "npm run build:qq -- --watch",
+    "dev:jd": "npm run build:jd -- --watch",
+    "dev:quickapp": "npm run build:quickapp -- --watch",
+    "temp": "node template"
+  },
+  "browserslist": [
+    "last 3 versions",
+    "Android >= 4.1",
+    "ios >= 8"
+  ],
+  "author": "",
+  "license": "MIT",
+  "dependencies": {
+    "@babel/runtime": "^7.7.7",
+    "@tarojs/components": "3.0.15",
+    "@tarojs/react": "3.0.15",
+    "@tarojs/runtime": "3.0.15",
+    "@tarojs/taro": "3.0.15",
+    "@types/classnames": "^2.2.11",
+    "classnames": "^2.2.6",
+    "hox": "^1.1.2",
+    "react": "^16.10.0",
+    "react-dom": "^16.10.0",
+    "react-redux": "^7.2.0",
+    "redux": "^4.0.0",
+    "redux-logger": "^3.0.6",
+    "redux-thunk": "^2.3.0"
+  },
+  "devDependencies": {
+    "@babel/core": "^7.8.0",
+    "@tarojs/mini-runner": "3.0.15",
+    "@tarojs/webpack-runner": "3.0.15",
+    "@types/react": "^16.0.0",
+    "@types/webpack-env": "^1.13.6",
+    "@typescript-eslint/eslint-plugin": "^2.x",
+    "@typescript-eslint/parser": "^2.x",
+    "babel-preset-taro": "3.0.15",
+    "eslint": "^6.8.0",
+    "eslint-config-taro": "3.0.15",
+    "eslint-plugin-import": "^2.12.0",
+    "eslint-plugin-react": "^7.8.2",
+    "eslint-plugin-react-hooks": "^4.2.0",
+    "prettier": "2.2.1",
+    "stylelint": "9.3.0",
+    "typescript": "^3.7.0"
+  }
 }
