@@ -1,105 +1,57 @@
 /**
- * pages模版快速生成脚本,执行命令 npm run tep `文件名`
- */
+ * pages模版快速生成脚本
+ **/
 
-const fs = require('fs');
+import FS from 'fs'
 
-const dirName = process.argv[2];
+const dirName = process.argv[2] //path
 
 if (!dirName) {
-  console.log('文件夹名称不能为空！');
-  console.log('示例：npm run tep test');
-  process.exit(0);
+  console.log('文件夹名称不能为空')
+  console.log('示例:npm run temp test')
+  process.exit(0)
 }
 
 // 页面模版
-const indexTep = `import Taro, { Component } from '@tarojs/taro';
-import { View } from '@tarojs/components';
-import { connect } from '@tarojs/redux';
-import './index.scss';
+const indexTemp = `import React from 'react'
+import Taro from '@tarojs/taro'
+import { View } from '@tarojs/components'
 
-@connect(({${dirName}}) => ({
-  ...${dirName},
-}))
-export default class ${titleCase(dirName)} extends Component {
-  config = {
-    navigationBarTitleText: '${dirName}',
-  };
+import './index.scss'
 
-  componentDidMount = () => {
-
-  };
-
-  render() {
-    return (
-      <View className="${dirName}-page">
-        ${dirName}
-      </View>
-    )
-  }
+const ${titleCase(dirName)}: Taro.FC = () => {
+  return (
+    <View className="index">
+      <View>Hello World</View>
+    </View>
+  )
 }
+export default ${titleCase(dirName)}
+
 `;
 
 // scss文件模版
-const scssTep = `@import "../../styles/mixin";
+const scssTemp = `@import '../../base.scss';
 
-.${dirName}-page {
-  @include wh(100%, 100%);
+.index {
+  height:100%;
+  @include flex-column;
 }
 `;
 
-// model文件模版
-const modelTep = `import * as ${dirName}Api from './service';
+//config 模板
+const configTemp = `export default {
+    navigationBarTitleText: '首页'
+  }
+  `
 
-export default {
-  namespace: '${dirName}',
-  state: {
+  
+FS.mkdirSync(`./src/pages/${dirName}`)
+process.chdir(`./src/pages/${dirName}`)
 
-  },
-
-  effects: {
-    * effectsDemo(_, { call, put }) {
-      const { status, data } = yield call(${dirName}Api.demo, {});
-      if (status === 'ok') {
-        yield put({ type: 'save',
-          payload: {
-            topData: data,
-          } });
-      }
-    },
-  },
-
-  reducers: {
-    save(state, { payload }) {
-      return { ...state, ...payload };
-    },
-  },
-
-};
-`;
-
-
-// service页面模版
-const serviceTep = `import Request from '../../utils/request';
-
-export const demo = data => Request({
-  url: '路径',
-  method: 'POST',
-  data,
-});
-`;
-
-
-
-fs.mkdirSync(`./src/pages/${dirName}`); // mkdir $1
-process.chdir(`./src/pages/${dirName}`); // cd $1
-
-fs.writeFileSync('index.js', indexTep);
-fs.writeFileSync('index.scss', scssTep);
-fs.writeFileSync('model.js', modelTep);
-fs.writeFileSync('service.js', serviceTep);
-
-console.log(`模版${dirName}已创建,请手动增加models`);
+FS.writeFileSync('index.tsx', indexTemp)
+FS.writeFileSync('index.scss', scssTemp)
+FS.writeFileSync('index.config.ts', configTemp)
 
 function titleCase(str) {
   const array = str.toLowerCase().split(' ');
@@ -107,7 +59,7 @@ function titleCase(str) {
     array[i] = array[i][0].toUpperCase() + array[i].substring(1, array[i].length);
   }
   const string = array.join(' ');
-  return string;
+  return string
 }
 
-process.exit(0);
+process.exit(0)
