@@ -4,13 +4,6 @@ module.exports = {
   },
   defineConstants: {
   },
-  weapp: {},
-  h5: {
-    devServer: {
-      host: 'localhost', // 如需局域网（如手机）访问，请更换为0.0.0.0
-      // host: '0.0.0.0', // 如需局域网（如手机）访问，请更换为0.0.0.0
-      port: 8088,
-      https: false
-    }
-  }
+  mini: {},
+  h5: {}
 }
