@@ -1,49 +1,61 @@
 const config = {
-  projectName: 'taro-msparis',
-  date: '2018-9-27',
-  // 设计稿尺寸
+  projectName: 'mymarket',
+  date: '2020-12-14',
   designWidth: 750,
+  deviceRatio: {
+    640: 2.34 / 2,
+    750: 1,
+    828: 1.81 / 2
+  },
   sourceRoot: 'src',
   outputRoot: 'dist',
-  // 通用插件配置
-  plugins: {
-    babel: {
-      sourceMap: true,
-      presets: [
-        'env'
-      ],
-      plugins: [
-        'transform-class-properties',
-        'transform-decorators-legacy',
-        'transform-object-rest-spread'
-      ]
-    },
+  plugins: [],
+  defineConstants: {
   },
-  // 全局变量设置
-  defineConstants: {},
-  // 小程序端专用配置
-  weapp: {
-    module: {
-      postcss: {
-        autoprefixer: {
-          enable: true
-        },
-        // 小程序端样式引用本地资源内联配置
-        url: {
-          enable: true,
-          limit: 10240
+  copy: {
+    patterns: [
+    ],
+    options: {
+    }
+  },
+  framework: 'react',
+  mini: {
+    postcss: {
+      pxtransform: {
+        enable: true,
+        config: {
+
+        }
+      },
+      url: {
+        enable: true,
+        config: {
+          limit: 1024 // 设定转换尺寸上限
+        }
+      },
+      cssModules: {
+        enable: false, // 默认为 false，如需使用 css modules 功能，则设为 true
+        config: {
+          namingPattern: 'module', // 转换模式，取值为 global/module
+          generateScopedName: '[name]__[local]___[hash:base64:5]'
         }
       }
     }
   },
-  // H5 端专用配置
   h5: {
     publicPath: '/',
     staticDirectory: 'static',
-    module: {
-      postcss: {
-        autoprefixer: {
-          enable: true
+    postcss: {
+      autoprefixer: {
+        enable: true,
+        config: {
+        }
+      },
+      cssModules: {
+        enable: false, // 默认为 false，如需使用 css modules 功能，则设为 true
+        config: {
+          namingPattern: 'module', // 转换模式，取值为 global/module
+          generateScopedName: '[name]__[local]___[hash:base64:5]'
         }
       }
     }
